Create nnLabeler_GPU.py
This is a GPU-enabled version of nnLabeler. If you have a moderately powerful, comtemporary GPU, training should be much faster than with nnLabeler.py (even when using multi-core enabled numpy on an i7 quad-core.)

The drawback is that you need to take time to setup CUDA and install the python packages 'cudamat', and 'gnumpy'.